fix: correct PostgreSQL type mismatches causing user list load failure
- users.rs: change is_active tuple type from i64 to i32 (PostgreSQL INTEGER = int4)
- app_catalog_cache.rs: fix app_id type i64 → i32 to match INTEGER column
- query_log_writer.rs + handler.rs: fix app_id Option<i64> → Option<i32>
- upstreams.rs: use SQL casts for BOOLEAN (::bigint) and TIMESTAMP (::text) columns
- api/mod.rs: fix health check query WHERE is_active = true and TIMESTAMP cast

Root cause: dns_upstreams uses PostgreSQL-native BOOLEAN/TIMESTAMP while other
tables use SQLite-compatible INTEGER/TEXT. SQLx rejects int4→i64 and bool→i64
decoding without explicit casts.

+++ src/api/handlers/users.rs
@@ -37,7 +37,7 @@ fn validate_role(role: &str) -> AppResult<()> {
 }
 
 pub async fn list(State(state): State<Arc<AppState>>, _admin: AdminUser) -> AppResult<Json<Value>> {
-    let rows: Vec<(String, String, String, i64, String, String)> = sqlx::query_as(
+    let rows: Vec<(String, String, String, i32, String, String)> = sqlx::query_as(
         "SELECT id, username, role, is_active, created_at, updated_at
          FROM users ORDER BY created_at DESC",
     )
@@ -145,7 +145,7 @@ pub async fn update_role(
     validate_role(&body.role)?;
 
     // Check if user exists
-    let existing: Option<(String, String, String, i64, String, String)> = sqlx::query_as(
+    let existing: Option<(String, String, String, i32, String, String)> = sqlx::query_as(
         "SELECT id, username, role, is_active, created_at, updated_at
          FROM users WHERE id = $1",
     )
